nav_types: make btin_t not packed
The use of bitfields of 10 bits makes the use of `unsigned char` impossible.
That means the compiler might use more bits than expected for the 4*12 bits.
This is the case with MS-compatible compilers. We cannot guarantee that the
btin_t will use the same 10 octets as used in the binary blob.

In the end btin_t is not read as a packed structure but bit by bit in
navRead_PCI().